update daily batch to use Polygon
