+++ dags/etl/load_daily_stock_price_dag.py
@@ -11,20 +11,21 @@ catalog_name = Variable.get("CATALOG_NAME")
 aws_region = Variable.get("AWS_GLUE_REGION")
 aws_access_key_id = Variable.get("DATAEXPERT_AWS_ACCESS_KEY_ID")
 aws_secret_access_key = Variable.get("DATAEXPERT_AWS_SECRET_ACCESS_KEY")
-staging_incremental_script_path = "jobs/batch/daily_stock_price_staging_incremental.py"
+staging_incremental_script_path = "jobs/batch/load_staging_daily_stock_price.py"
 
-# #Alpaca keys
-apca_api_key_id = Variable.get("APCA_API_KEY_ID")
-apca_api_secret_key = Variable.get("APCA_API_SECRET_KEY")
+# # #Alpaca keys
+# apca_api_key_id = Variable.get("APCA_API_KEY_ID")
+# apca_api_secret_key = Variable.get("APCA_API_SECRET_KEY")
+
+#Polygon key
+polygon_api_key = Variable.get("POLYGON_API_KEY")
 
 
-
-
-@dag("daily_stock_price_incremental_dag",
+@dag("load_daily_stock_price_dag",
      description="Load the previous day's stock data to staging, perform quality checks, and publish",
      default_args={
          "owner": "William Switzer",
-         "start_date": datetime(2024, 6, 21),
+         "start_date": datetime(2024, 6, 22),
          "retries": 1,
      },
      max_active_runs=1,
@@ -32,17 +33,17 @@ apca_api_secret_key = Variable.get("APCA_API_SECRET_KEY")
      catchup=True,
      tags=["pyspark", "glue", "eczachly", "billyswitzer"],
      template_searchpath='jobs')
-def daily_stock_price_incremental_dag():
-    staging_incremental_flat_table = "billyswitzer.staging_daily_stock_price_incremental"
-    staging_incremental_cumulative_table = "billyswitzer.staging_daily_stock_price_cumulative"
+def load_daily_stock_price_dag():
+    staging_daily_flat_table = "billyswitzer.staging_daily_stock_price"
+    staging_daily_cumulative_table = "billyswitzer.staging_daily_stock_price_cumulative"
     production_cumulative_table = "billyswitzer.daily_stock_price_cumulative"
     production_dim_table = "billyswitzer.dim_daily_stock_price"
 
-    load_staging_incremental_flat_table = PythonOperator(
-        task_id="load_staging_incremental_flat_table",
+    load_staging_flat_table = PythonOperator(
+        task_id="load_staging_flat_table",
         python_callable=create_glue_job,
         op_kwargs={
-            "job_name": "switzer_daily_stock_price_incremental_job",
+            "job_name": "switzer-staging_daily_stock_price_job",
             "script_path": staging_incremental_script_path,
             "aws_access_key_id": aws_access_key_id,
             "aws_secret_access_key": aws_secret_access_key,
@@ -50,12 +51,11 @@ def daily_stock_price_incremental_dag():
             "s3_bucket": s3_bucket,
             "catalog_name": catalog_name,
             "aws_region": aws_region,
-            "description": "Incremental Load to staging table",
+            "description": "Daily load to staging table",
             "arguments": {
                 "--ds": "{{ ds }}",
-                "--output_table": staging_incremental_flat_table,
-                "--apca_api_key_id": apca_api_key_id,
-                "--apca_api_secret_key": apca_api_secret_key
+                "--output_table": staging_daily_flat_table,
+                '--polygon_api_key': polygon_api_key
             },
         },
     )
@@ -68,51 +68,33 @@ def daily_stock_price_incremental_dag():
                 SELECT COUNT(CASE WHEN close_price IS NULL THEN 1 END) = 0 AS close_price_is_not_null_check,
                     COUNT(CASE WHEN high_price IS NULL THEN 1 END) = 0 AS high_price_is_not_null_check,
                     COUNT(CASE WHEN low_price IS NULL THEN 1 END) = 0 AS low_price_is_not_null_check,
-                    COUNT(CASE WHEN trade_count IS NULL THEN 1 END) = 0 AS trade_count_is_not_null_check,
                     COUNT(CASE WHEN open_price IS NULL THEN 1 END) = 0 AS open_price_is_not_null_check,
                     COUNT(CASE WHEN volume IS NULL THEN 1 END) = 0 AS volume_is_not_null_check,
-                    COUNT(CASE WHEN volume_weighted_average_price IS NULL THEN 1 END) = 0 AS volume_weighted_average_price_is_not_null_check,
                     COUNT(CASE WHEN symbol IS NULL THEN 1 END) = 0 AS symbol_is_not_null_check,
                     COUNT(CASE WHEN as_of_date IS NULL THEN 1 END) = 0 AS as_of_date_is_not_null_check
-                FROM {staging_incremental_flat_table}
+                FROM {staging_daily_flat_table}
             """
         }
     )
 
-    cleanup_staging_incremental_cumulative_table_pre = PythonOperator(
-        task_id="cleanup_staging_incremental_cumulative_table_pre",
+    cleanup_staging_cumulative_table_pre = PythonOperator(
+        task_id="cleanup_staging_cumulative_table_pre",
         python_callable=execute_trino_query,
         op_kwargs={
-            'query': f"""DELETE FROM {staging_incremental_cumulative_table}"""
+            'query': f"""DELETE FROM {staging_daily_cumulative_table}"""
         }
-    )    
+    )
 
     stage_cumulative_table_step = PythonOperator(
         task_id="stage_cumulative_table_step",
         python_callable=execute_trino_query,
         op_kwargs={
             'query': f"""
-                    INSERT INTO {staging_incremental_cumulative_table}
-                    WITH date_cte AS 
-                    (
-                        SELECT DATE('{{{{ ds }}}}') AS current_partition_date,
-                        DATE('{{{{ ds }}}}') - INTERVAL '1' DAY AS last_partition_date
-                    ),
-                    yesterday AS 
-                    (
-                    SELECT 
-                        symbol,
-                        as_of_date,
-                        FILTER(price_array, x -> x.bar_date > as_of_date - INTERVAL '364' DAY) AS price_array
-                    FROM {production_cumulative_table} p
-                        JOIN date_cte dc ON p.as_of_date = dc.last_partition_date
-                    ),
-                    today AS 
-                    (
+                    INSERT INTO {staging_daily_cumulative_table}
                     SELECT symbol,
-                        ARRAY_AGG(ROW(close_price, high_price, low_price, trade_count, open_price, bar_date, volume, volume_weighted_average_price)) AS price_array,
+                        ARRAY_AGG(ROW(close_price, high_price, low_price, trade_count, open_price, bar_date, volume, volume_weighted_average_price) ORDER BY bar_date DESC) AS price_array,
                         as_of_date
-                    FROM {staging_incremental_flat_table}
+                    FROM {staging_daily_flat_table}
                     WHERE close_price > 0
                         AND high_price > 0
                         AND low_price > 0
@@ -122,22 +104,6 @@ def daily_stock_price_incremental_dag():
                         AND volume_weighted_average_price > 0
                     GROUP BY symbol,
                         as_of_date
-                    ),
-                    combined AS 
-                    (
-                        SELECT COALESCE(y.symbol, t.symbol) AS symbol,
-                        CASE WHEN y.price_array IS NULL THEN t.price_array
-                            WHEN t.price_array IS NULL THEN y.price_array
-                            ELSE t.price_array || y.price_array END AS price_array
-                        FROM yesterday y
-                        FULL OUTER JOIN today t on y.symbol = t.symbol
-                            AND y.as_of_date + INTERVAL '1' DAY = t.as_of_date
-                    )
-                    SELECT symbol,
-                        price_array,
-                        current_partition_date
-                    FROM combined c
-                        JOIN date_cte dc ON 1=1
                """
         }
     )
@@ -160,7 +126,7 @@ def daily_stock_price_incremental_dag():
                 staging_table AS
                 (
                     SELECT COUNT(1) AS row_count
-                    FROM {staging_incremental_cumulative_table}
+                    FROM {staging_daily_cumulative_table}
                 )
                 SELECT st.row_count < pt.row_count * 1.01 AS new_rows_under_threshold_check  
                 FROM production_table pt
@@ -184,7 +150,7 @@ def daily_stock_price_incremental_dag():
         op_kwargs={
             'query': f"""
                    INSERT INTO {production_cumulative_table}
-                   SELECT * FROM {staging_incremental_cumulative_table}
+                   SELECT * FROM {staging_daily_cumulative_table}
                """
         }
     )
@@ -193,7 +159,7 @@ def daily_stock_price_incremental_dag():
         task_id="cleanup_staging_incremental_cumulative_table_post",
         python_callable=execute_trino_query,
         op_kwargs={
-            'query': f"""DELETE FROM {staging_incremental_cumulative_table}"""
+            'query': f"""DELETE FROM {staging_daily_cumulative_table}"""
         }
     )
 
@@ -218,9 +184,9 @@ def daily_stock_price_incremental_dag():
                 ), previous_weekday AS
                 (
                 SELECT CASE
-                    WHEN DAY_OF_WEEK(current_partition_date) = 7 THEN current_partition_date - INTERVAL '2' DAY		--Sunday
-                    WHEN DAY_OF_WEEK(current_partition_date) = 6 THEN current_partition_date - INTERVAL '1' DAY		--Saturday
-                    ELSE current_partition_date
+                        WHEN DAY_OF_WEEK(current_partition_date) = 7 THEN current_partition_date - INTERVAL '2' DAY		--Sunday
+                        WHEN DAY_OF_WEEK(current_partition_date) = 6 THEN current_partition_date - INTERVAL '1' DAY		--Saturday
+                        ELSE current_partition_date
                     END AS prev_weekday
                 FROM partition_date_cte 
                 ),
@@ -251,9 +217,9 @@ def daily_stock_price_incremental_dag():
         }
     )    
 
-    load_staging_incremental_flat_table >> \
+    load_staging_flat_table >> \
         run_dq_not_null_flat_table_check >> \
-            cleanup_staging_incremental_cumulative_table_pre >> \
+            cleanup_staging_cumulative_table_pre >> \
                 stage_cumulative_table_step >> \
                     run_dq_new_rows_under_threshold_check >> \
                         delete_current_partition_cumulative_prod >> \
@@ -263,5 +229,5 @@ def daily_stock_price_incremental_dag():
                                         load_dim_daily_stock_price_table
 
 
-daily_stock_price_incremental_dag()
+load_daily_stock_price_dag()
